Replace lists of array indices with tuples
This removes the FutureWarning from NumPy.

The change in NumPy was probably made to avoid confusion between lists
being used for adding an extra dimension, as to grouping
arrays(*). Using a tuple makes this less ambiguous. If the combined
arrays were to be combined into a multi-dimensional index array,
functions like `tile` or `vstack` should be used.

(*) The warning itself states: "In the future [using a list] will be
interpreted as an array index, `arr[np.array(seq)]`, which will result
either in an error or a different result."